refactor: token_minter test cleanup
