+++ arbiter-engine/src/examples/minter/behaviors/mod.rs
@@ -1,0 +1,3 @@
+use super::*;
+pub mod token_admin;
+pub mod token_requester;
